03 - APIs FastAPI/02 - Introducao ao FastAPI/dio_fastapi/pyproject.toml
[project]
name = "dio-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Samuel-Main",email = "patriciosamuel097@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.120.0,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "databases[aiosqlite] (>=0.9.0,<0.10.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
